[WIP] Started working on TextArea aggregate widget, (...) - Don't clear pointer style for wheel events - Use LayoutContraints instead of unlabelled 4-tuple for Root.constraints accessor - Added ScrollbarMode.Hidden   - Allows content to be scrolled, but scrollbars are hidden, don't receive pointer events, and don't reserve space when auto-scrolling - Added workaround for blurry caret, jittery text and TextInput panning due to error between idealWidth and width   - Fix does not work for text that would require non-integer offsets, such as typing text and auto-scrolling along - Fixed panning jittering if the height was smaller than the full line height (line height + line spacing)
TODO: There are too many text-alignment issues that were uncovered by the TextArea widget, which were introduced with the updated that fixed clipping. It's better to have awkward text alignment and perfect clipping instead of perfect text alignment and broken clipping, but there should be a way to have both. How?